requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.1
openai==1.30.0
pytest==8.2.0
httpx[http2]==0.27.0
bcrypt==4.1.3
alembic==1.13.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0